fix: minor classname tweeks
